requirements.txt
certifi==2020.12.5
click==7.1.2
joblib==1.0.1
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
python-dateutil==2.8.1
pytorch-crf==0.7.2
pytz==2021.1
regex==2021.3.17
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
torch==1.8.0
torchvision==0.9.0
tqdm==4.59.0
typing-extensions==3.7.4.3
